Move StaticNode into the nodes submodule
