+++ pyiron_workflow/nodes/function.py
@@ -6,7 +6,8 @@ from typing import Any
 from pyiron_snippets.colors import SeabornColors
 from pyiron_snippets.factory import classfactory
 
-from pyiron_workflow.io_preview import StaticNode, ScrapesIO
+from pyiron_workflow.io_preview import ScrapesIO
+from pyiron_workflow.nodes.static_io import StaticNode
 
 
 class Function(StaticNode, ScrapesIO, ABC):
